L3-IPv4-ACL.yml: New playbook.
This playbook is designed to push an entire access-list to a device.
If the access-list is already present, it will be removed before
applying the new access-list.

To help prevent a lockout, the playbook will check if the access-list
has been configured as the SSH access-group.  If the access-list is
configured as the SSH access-group, the playbook will first check if
the IP address of the AWX server is permitted in the access-list and
the playbook will force an error if it is denied. If the IP address is
permitted, the SSH access-group command will be removed befored applying
the access-list config, then re-added after.